+++ priv/repo/migrations/20171107074327_test_migration.exs
@@ -1,0 +1,11 @@
+defmodule Guardian.OneTime.Repo.Migrations.TestMigration do
+  use Ecto.Migration
+
+  def change do
+    create table(:one_time_tokens, primary_key: false) do
+      add :id, :string, priary_key: true
+      add :claims, :json
+      add :expiry, :utc_datetime
+    end
+  end
+end